requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
backports.zoneinfo;python_version<"3.9"
boto3==1.24.22
botocore==1.27.22
Django==4.0.5
django-cors-headers==3.13.0
django-mysql==4.7.0
django-storages==1.12.3
django-tables2==2.4.1
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
requests~=2.28.1

# numpy~=1.23.0
# pandas~=1.4.3
# deepface~=0.0.75
# Pillow~=9.2.0
# opencv-python-headless

numpy~=1.23.0
Pillow~=9.2.0
django-crispy-forms~=1.14.0
PyQRCode~=1.2.1
luxand~=0.7
datefinder~=0.7.1